# argos-service-api/api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: Apache License 2.0
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
  /supplychain:
    post:
      summary: create a supplychain
      operationId: createSupplyChain
      tags:
        - supplychain
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '201':
          description: created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created supplychain can be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get supply chain by path
      operationId: getSupplyChainByPath
      tags:
        - supplychain
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          style: simple
          explode: true
          required: true
          allowEmptyValue: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: the supply chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '404':
          description: not found
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}:
    get:
      summary: get a supplychain
      operationId: getSupplyChain
      tags:
        - supplychain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: the supply chain item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '500':
          $ref: '#/components/responses/Error'
    put:
      summary: update supplychain by id
      operationId: updateSupplyChain
      tags:
        - supplychain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '200':
          description: updated supply chain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyChain'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
...
  /serviceaccount:
    post:
      summary: create a new service account
      operationId: createServiceAccount
      tags:
        - service account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccount"
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created service account can be found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /serviceaccount/me/activekey:
    get:
      summary: return the active key pair
      operationId: getServiceAccountKey
      security:
        - basicAuth: []
      tags:
        - service account
      responses:
        '200':
          description: active key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountKeyPair"
        '404':
          description: no active key exists
        '500':
          $ref: '#/components/responses/Error'
  /serviceaccount/{serviceAccountId}:
    get:
      summary: find service account by id
      operationId: getServiceAccountById
      tags:
        - service account
      parameters:
        - name: serviceAccountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: service account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        '500':
          $ref: '#/components/responses/Error'
    put:
      summary: update service account by id
      operationId: updateServiceAccountById
      tags:
        - service account
      parameters:
        - name: serviceAccountId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccount"
      responses:
        '200':
          description: updated service account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /serviceaccount/{serviceAccountId}/key:
    post:
      summary: create a new active key the old key will be made inactive
      operationId: createServiceAccountKeyById
      tags:
        - service account
      parameters:
        - name: serviceAccountId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountKeyPair"
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountKeyPair"
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created service account can be found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: return the active key pair
      operationId: getServiceAccountKeyById
      tags:
        - service account
      parameters:
        - name: serviceAccountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountKeyPair"
        '404':
          description: no active key exists
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    ServiceAccountKeyPair:
      type: object
      required:
        - keyId
        - publicKey
        - hashedKeyPassphrase
        - encryptedPrivateKey
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          type: string
          format: byte
        hashedKeyPassphrase:
          type: string
          writeOnly: true
          description: used for authentication of a service account via http basic authentication
        encryptedPrivateKey:
          type: string
          format: byte
    SupplyChain:
      type: object
      required:
        - name
        - parentLabelId
      properties:
        id:
          type: string
          maxLength: 36
          readOnly: true
        name:
          type: string
          maxLength: 256
          pattern: '^([a-z]{1}[a-z0-9-]*)?$'
        parentLabelId:
          type: string
          minLength: 36
          maxLength: 36
          pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}?$'
...
    KeyId:
      type: string
      description: key id is a hash of the public key
      minLength: 64
      maxLength: 64
      pattern: '^[0-9a-f]*$'
...
    ServiceAccount:
      type: object
      required:
        - name
        - parentLabelId
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
        parentLabelId:
          type: string
          minLength: 36
          maxLength: 36
          pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}?$'
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          type: string
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string